+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Improvements in verification of Audit consistency proofs
+
 ## [3.9.0] - 2024-06-07
 
 ### Added